Revert "feat: create and implement `Event` class (#6)"
This reverts commit b256b1caf052928a06452eedbdca76e702abadd5.
