+++ .gitignore
@@ -8,7 +8,3 @@ build/
 # Omit committing pubspec.lock for library packages; see
 # https://dart.dev/guides/libraries/private-files#pubspeclock.
 pubspec.lock
-
-# Omit generated files.
-# See https://dart.dev/guides/libraries/private-files#generated-files.
-*.g.dart
