--- lib/error.dart
@@ -1,3 +1,0 @@
-library isen_aurion_client;
-
-export 'src/error.dart';
