--- lib/event.dart
@@ -1,3 +1,0 @@
-library isen_aurion_client;
-
-export 'src/event.dart';
