+++ lib/isen_aurion_client.dart
@@ -1,0 +1,9 @@
+/// Support for doing something awesome.
+///
+/// More dartdocs go here.
+library isen_aurion_client;
+
+export 'src/aurion.dart';
+export 'src/error.dart';
+
+// TODO: Export any libraries intended for clients of this package.
